ipmi tables: extract IPMIClient into own file & other minor changes
